requirements.txt
Package             Version  
------------------- ---------
certifi             2018.8.24
chardet             3.0.4    
defusedxml          0.5.0    
Django              2.2.5    
django-allauth      0.37.1   
django-cors-headers 2.4.0    
django-rest-auth    0.9.3    
djangorestframework 3.8.2    
gunicorn            19.7.1   
httpie              1.0.3    
idna                2.7      
oauthlib            2.1.0    
pip                 19.1.1   
Pygments            2.4.2    
python3-openid      3.1.0    
pytz                2018.5   
requests            2.22.0   
requests-oauthlib   1.0.0    
setuptools          41.0.1   
six                 1.11.0   
sqlparse            0.3.0    
urllib3             1.23     
wheel               0.33.4   
whitenoise          3.3.1    
